smart-contract/requirements.txt
aiodns==3.0.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
ape-alchemy==0.8.7
ape-ens==0.8.1
ape-etherscan==0.8.4
ape-foundry==0.8.7
ape-hardhat==0.8.5
ape-infura==0.8.4
ape-solidity==0.8.5
ape-template==0.8.0
ape-tokens==0.8.3
ape-vyper==0.8.8
argcomplete==2.0.0
arrow==1.3.0
asttokens==2.4.1
async-timeout==5.0.1
attrs==24.3.0
awscli==2.15.2
awscrt==0.22.4
Babel==2.12.1
base58==1.0.3
bcrypt==4.0.1
Beaker==1.12.1
beautifulsoup4==4.12.3
bidict==0.23.1
binaryornot==0.4.4
bitarray==3.0.0
blivet==3.8.2
blivet-gui==2.6.0
Brlapi==0.8.5
Brotli==1.1.0
cached-property==2.0.1
cbor2==5.6.5
cchecksum==0.0.4
certifi==2024.12.14
cffi==1.15.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.1
chompjs==1.3.0
ckzg==2.0.1
click==8.1.8
click-plugins==1.1.1
colorama==0.4.6
cookiecutter==2.1.1
cryptography==41.0.7
cupshelpers==1.0
cytoolz==1.0.1
dasbus==1.7
dataclassy==0.11.1
dbus-python==1.3.2
decorator==5.1.1
distlib==0.3.6
distro==1.8.0
dnf==4.21.1
docutils==0.19
ecdsa==0.19.0
eip712==0.2.10
eth-account==0.13.4
eth-ape==0.8.24
eth-bloom==3.1.0
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-pydantic-types==0.1.3
eth-rlp==2.1.0
eth-tester==0.12.0b2
eth-typing==5.1.0
eth-utils==5.2.0
eth_abi==5.2.0
ethpm-types==0.6.19
evm-trace==0.2.4
evmchains==0.1.3
exceptiongroup==1.2.2
executing==2.1.0
fedora-third-party==0.10
file-magic==0.4.0
filelock==3.12.2
fros==1.1
frozenlist==1.5.0
gevent==24.11.1
greenlet==3.1.1
h11==0.14.0
hexbytes==1.3.0
httptools==0.6.4
humanize==3.13.1
identify==2.5.36
idna==3.10
ijson==3.3.0
importlib_metadata==8.6.1
iniconfig==2.0.0
ipython==8.31.0
jaraco.classes==3.3.0
jedi==0.19.2
jeepney==0.8.0
Jinja2==3.1.5
jinja2-time==0.2.0
jmespath==1.0.1
keyring==23.11.0
langtable==0.0.68
lazyasd==0.1.4
libcomps==0.1.20
libdnf==0.73.3
lru-dict==1.3.0
lxml==4.9.3
Mako==1.2.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==8.13.0
morphys==1.0
msgspec==0.19.0
multidict==6.1.0
mutagen==1.46.0
nftables==0.1
nodeenv==1.9.1
numpy==1.26.4
nvautoinstall==0.4.2
olefile==0.46
packaging==23.2
pandas==2.2.3
parsimonious==0.10.0
parso==0.8.4
Paste==3.5.3
pexpect==4.9.0
pid==2.2.3
pillow==10.3.0
platformdirs==3.9.1
pluggy==1.5.0
ply==3.11
pre-commit==3.8.0
productmd==1.38
progressbar2==3.53.2
prompt_toolkit==3.0.50
propcache==0.2.1
proton-core==0.3.3
proton-keyring-linux==0.1.0
proton-vpn-api-core==0.35.5
proton-vpn-gtk-app==4.6.0
proton-vpn-network-manager==0.9.1
ptyprocess==0.7.0
pure_eval==0.2.3
pwquality==1.4.5
py-cid==0.3.0
py-ecc==7.0.1
py-evm==0.10.1b2
py-geth==5.2.0
py-multibase==1.0.3
py-multicodec==0.2.1
py-multihash==0.2.3
py-solc-x==2.0.3
pyasn1==0.6.1
pycairo==1.25.1
pycares==4.3.0
pycparser==2.20
pycrypto==2.6.1
pycryptodome==3.21.0
pycryptodomex==3.21.0
pycups==2.0.4
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
pyenchant==3.2.2
pygit2==1.14.0
Pygments==2.19.1
PyGObject==3.46.0
pykickstart==3.48
PyMySQL==1.1.1
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pyparted==3.13.0
PySocks==1.7.1
pytest==8.3.4
python-augeas==1.1.0
python-baseconv==1.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.10.1
python-engineio-client==1.1
python-gnupg==0.5.0
python-meh==0.51
python-slugify==8.0.4
python-utils==3.3.3
pytz==2024.2
pyudev==0.24.1
pyunormalize==16.0.0
pyxdg==0.27
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-file==1.5.1
requests-ftp==0.3.1
rich==13.9.4
rlp==4.0.1
rpm==4.19.1.1
rpmautospec==0.7.3
rpmautospec-core==0.1.5
rsa==4.9
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
safe-pysha3==1.0.4
SecretStorage==3.3.3
selinux @ file:///builddir/build/BUILD/libselinux-3.5/src
semantic-version==2.10.0
sentry-sdk==2.13.0
sepolicy @ file:///builddir/build/BUILD/selinux-3.5/python/sepolicy
setools==4.4.3
setuptools==67.7.2
simple-websocket==1.1.0
simpleaudio==1.0.4
simpleline==1.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sos==4.6.0
soupsieve==2.5
SQLAlchemy==2.0.37
sqlparse==0.5.1
stack-data==0.6.3
starlette==0.40.0
systemd-python==235
Tempita==0.5.2
text-unidecode==1.3
tokenlists==0.1.7
toml==0.10.2
tomli==2.2.1
toolz==1.0.0
tqdm==4.67.1
traitlets==5.14.3
trie==3.0.1
types-python-dateutil==2.9.0.20241206
types-requests==2.32.0.20241016
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvloop==0.21.0
varint==1.0.2
virtualenv==20.21.1
vvm==0.3.2
vyper==0.4.0
watchdog==3.0.0
watchfiles==1.0.0
wcwidth==0.2.13
web3==7.7.0
websockets==13.1
wheel==0.45.1
wsproto==1.2.0
yarl==1.18.3
yt-dlp==2024.8.6
zipp==3.21.0
zope.event==5.0
zope.interface==7.2